requirements.txt
asgiref==3.3.4
CacheControl==0.12.6
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
charset-normalizer==2.0.4
cryptography==3.4.7
cycler==0.10.0
Django==3.2.6
firebase-admin==4.5.0
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-cloud-core==1.5.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.35.0
google-crc32c==1.1.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
httplib2==0.19.1
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.3
mod-wsgi==4.7.1
msgpack==1.0.2
mysql-connector==2.2.9
mysqlclient==2.0.3
numpy==1.19.5
Pillow==8.3.2
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
requests==2.26.0
rsa==4.7.2
six==1.15.0
sqlparse==0.4.2
uritemplate==3.0.1
urllib3==1.26.6
